feat(record): merge the in-meeting chat panel into the assistant card
One inline conversation thread replaces the two overlapping AI surfaces (the
Assistant card + the slide-out Chat panel). Voice and text now share ONE
chronological thread with multi-turn memory; the thread clears on each new
recording. Pure FE consolidation — no backend change, no version bump.

- AssistantStore is now the single store: a chronological messages thread
  (send() ships full history -> ask_assistant_chat memory), voice results
  (onWake/onResult) land in the same thread, init() subscribes to BOTH
  tool-trace streams; clear() added; one-shot interactions model removed.
- assistant-actions.component: a scrollable thread + input (mic + composer)
  pinned at the bottom, afterNextRender auto-scroll.
- record.component: removed the panel/FAB/showChat + MeetingChatStore; added a
  clear-on-record effect (allowSignalWrites, false->true edge).
- Deleted meeting-chat.store.ts + meeting-chat-panel.component.ts.

Spec: docs/superpowers/specs/2026-06-30-unified-assistant-merge-design.md
Verified: ng lint + ng build green; adversarial-verifier PASS (19/19 live
mocked-IPC assertions; NG0600 RED-before-GREEN).
