+++ src/app/core/assistant.store.ts
@@ -2,32 +2,28 @@ import { Injectable, computed, inject, signal } from "@angular/core";
 import type { UnlistenFn } from "@tauri-apps/api/event";
 import { IpcService } from "./ipc.service";
 import type {
+  AssistantToolPayload,
+  ChatMsg,
   VoiceActionResultPayload,
   VoiceActionStatus,
   VoiceCommandListeningPayload,
   VoiceCommandProcessingPayload,
-  AssistantToolPayload,
   WakeDetectedPayload,
 } from "./models";
 
 /**
  * The 4-state visual model of the assistant orb (industry-convergent
  * idle → listening → processing → answer). Pure presentation: derived from the
- * store's listening/processing/in-flight signals + the newest interaction
+ * store's listening/processing/in-flight signals + the newest assistant bubble
  * status by {@link AssistantStore.orbState}, never set directly.
  */
 export type OrbState = "idle" | "listening" | "processing" | "answer";
 
 /**
- * Phase H — one recent in-meeting voice-assistant interaction. A wake creates a
- * `pending` row ("heard: {command}"); the matching result resolves it with a
- * summary + citation chips + a status pill. Newest-first.
- */
-/**
  * One parsed grounding citation. The backend sends a flat `string[]` mixing two
  * shapes (`voice_action.rs`): a VAULT meeting wikilink `[[Title]]`, and a WEB hit
  * `(web) Title — https://…` (the loud "via web" attribution). We parse each into
- * this discriminated shape so the card renders vault chips and "via web" links
+ * this discriminated shape so the surface renders vault chips and "via web" links
  * distinctly — a web source is visibly off-device, never a `[[vault]]` chip.
  */
 export interface AssistantCitation {
@@ -41,7 +37,7 @@ export interface AssistantCitation {
 /**
  * One tool the brain used during an agentic turn — drives the LIVE trace chips
  * ("Searching notes… ✓", "Checking the web…"). Carries the tool name + a coarse
- * count only (no PII). Pushed/updated by {@link AssistantStore.onAssistantTool}.
+ * count only (no PII). Pushed/updated by {@link AssistantStore.onTool}.
  */
 export interface ToolTraceStep {
   /** Stable id for `@for` tracking (never key a trace chip on $index). */
@@ -56,20 +52,24 @@ export interface ToolTraceStep {
   count: number | null;
 }
 
-export interface AssistantInteraction {
-  /** Stable id for `@for` tracking (we never key on $index). */
+/**
+ * One message in the unified in-meeting assistant thread. Voice and text turns
+ * both land here as a `user` bubble (the heard/typed command) paired with an
+ * `assistant` bubble (resolved from "pending"), so speech and text share ONE
+ * chronological conversation. A user message carries the question; an assistant
+ * message carries the brain's answer, its live tool-trace, and its citations.
+ */
+export interface ChatMessage {
+  /** Stable id for `@for` tracking (never key on $index). */
   id: number;
-  /** Where the question came from — drives the row icon (mic / keyboard / wake). */
-  source: "voice" | "wake" | "text";
-  /** What the user said (voice) or typed (text). */
-  command: string;
-  /** "pending" until the result arrives, then mirrors the result status. */
+  role: "user" | "assistant";
+  /** The question (user) or the answer markdown (assistant; empty while pending). */
+  text: string;
+  /** Assistant only: "pending" while in flight, then the result status. */
   status: "pending" | VoiceActionStatus;
-  /** The assistant's answer (empty while pending). */
-  summary: string;
-  /** The live tool-trace for THIS turn (the chips), in call order. */
+  /** Assistant only: the live tool-trace chips for this turn. */
   trace: ToolTraceStep[];
-  /** Parsed grounding citations → vault `[[Title]]` chips + "via web" links. */
+  /** Assistant only: grounding citations (vault `[[Title]]` + "via web"). */
   citations: AssistantCitation[];
 }
 
@@ -101,47 +101,49 @@ export function parseCitations(raw: string[]): AssistantCitation[] {
 }
 
 /**
- * Signal-based store for the in-meeting voice assistant. Subscribes ONCE (the
- * RecorderStore.init() pattern) to the wake + result event streams and keeps a
- * capped, newest-first list of interactions. No NgRx, no subscribe-into-a-field:
- * the event payloads land in a `signal`.
+ * The SINGLE in-meeting assistant store: one chronological conversation thread
+ * fed by BOTH voice and text, plus the voice input-state (orb / listening /
+ * processing). Subscribes ONCE (the RecorderStore.init() pattern) to the wake +
+ * result + listening + processing + BOTH tool-trace streams and lands every
+ * payload in a `signal` — no NgRx, no subscribe-into-a-field.
+ *
+ * Memory: a TEXT `send` ships the FULL clean history to `ask_assistant_chat`
+ * (multi-turn memory); a VOICE turn answers one-shot via the voice backend and
+ * is appended to the SAME thread, so a typed follow-up remembers what was said.
  */
 @Injectable({ providedIn: "root" })
 export class AssistantStore {
   private readonly ipc = inject(IpcService);
 
-  /** Most recent N interactions kept; older ones drop off (no unbounded growth). */
-  private static readonly MAX = 12;
-
-  private readonly _interactions = signal<AssistantInteraction[]>([]);
-  /** Newest-first list of recent assistant interactions. */
-  readonly interactions = this._interactions.asReadonly();
-
-  /** Whether any interaction has ever been observed (drives empty-state copy). */
-  readonly hasAny = computed(() => this._interactions().length > 0);
+  private readonly _messages = signal<ChatMessage[]>([]);
+  /** The unified conversation, oldest → newest. */
+  readonly messages = this._messages.asReadonly();
+  /** Whether any turn exists yet (drives the empty-state copy). */
+  readonly hasMessages = computed(() => this._messages().length > 0);
 
   /**
    * True while the manual "Ask AI" listener has the mic open (between the
    * `{active:true}` and `{active:false}` EVENT_VOICE_COMMAND_LISTENING events).
-   * Drives the pulsing button + "🎙 Słucham…" inline indicator.
+   * Drives the pulsing mic button + "🎙 Słucham…" inline indicator.
    */
   private readonly _listening = signal(false);
   readonly listening = this._listening.asReadonly();
 
   /**
    * True from the instant a manual ask is fired until its answer lands — keeps
-   * the assistant-actions card visible across the whole round-trip even when the
-   * realtime-reactions config toggle is off. Set by {@link askStarted}, cleared
-   * when a result resolves a pending row.
+   * the assistant surface visible across the whole round-trip even when the
+   * realtime-reactions config toggle is off. Set by {@link askNow}, cleared when
+   * a result resolves a pending bubble.
    */
   private readonly _manualAskInFlight = signal(false);
   readonly manualAskInFlight = this._manualAskInFlight.asReadonly();
 
   /**
-   * True while a dispatched command is being processed (between the listener
-   * stopping and the answer landing) — the EVENT_VOICE_COMMAND_PROCESSING
-   * `{active}` boolean. Drives the orb's PROCESSING state + the "🧠 Przetwarzam…"
-   * shimmer label. Cleared by the result (or by the end-ask error path).
+   * True while a dispatched command (voice OR text) is being processed — the
+   * gap between the listener stopping / the text turn dispatching and the answer
+   * landing. Drives the orb's PROCESSING state, the composer disable, and the
+   * "🧠 Przetwarzam…" shimmer label. Cleared by the result (voice) or the
+   * send() finally (text), or by the end-ask error path.
    */
   private readonly _processing = signal(false);
   readonly processing = this._processing.asReadonly();
@@ -152,19 +154,22 @@ export class AssistantStore {
    *   processing → "processing" (highest priority: a dispatch is in flight)
    *   listening  → "listening"  (the mic is open)
    *   manual ask in flight (begin clicked, listener not yet open) → "listening"
-   *   newest interaction resolved → "answer" (a result is on screen)
+   *   newest assistant bubble resolved → "answer" (a result is on screen)
    *   otherwise → "idle".
-   * Bound on the orb as `[state]="orbState()"`.
    */
   readonly orbState = computed<OrbState>(() => {
     if (this._processing()) return "processing";
     if (this._listening() || this._manualAskInFlight()) return "listening";
-    const top = this._interactions()[0];
-    if (top && top.status !== "pending") return "answer";
+    const msgs = this._messages();
+    for (let i = msgs.length - 1; i >= 0; i--) {
+      if (msgs[i].role === "assistant") {
+        return msgs[i].status !== "pending" ? "answer" : "idle";
+      }
+    }
     return "idle";
   });
 
-  /** Monotonic id source for interaction rows (stable `@for` keys). */
+  /** Monotonic id source for message bubbles (stable `@for` keys). */
   private nextId = 1;
   /** Monotonic id source for tool-trace chips (stable `@for` keys). */
   private nextTraceId = 1;
@@ -174,12 +179,13 @@ export class AssistantStore {
   private unlistenListening: UnlistenFn | null = null;
   private unlistenProcessing: UnlistenFn | null = null;
   private unlistenTool: UnlistenFn | null = null;
+  private unlistenChatTool: UnlistenFn | null = null;
   /** Synchronous re-entrancy guard so two concurrent init() calls (e.g. the record
-   * screen + the card both initialising) can't double-subscribe before the first
+   * screen + the surface both initialising) can't double-subscribe before the first
    * `await` resolves. */
   private initializing = false;
 
-  /** Subscribe once to the wake + result streams. Idempotent + concurrency-safe. */
+  /** Subscribe once to the wake/result/listening/processing + both tool streams. */
   async init(): Promise<void> {
     if (this.unlistenWake || this.initializing) return;
     this.initializing = true;
@@ -193,9 +199,11 @@ export class AssistantStore {
     this.unlistenProcessing = await this.ipc.onVoiceCommandProcessing((p) =>
       this.onProcessing(p),
     );
-    this.unlistenTool = await this.ipc.onAssistantTool((p) =>
-      this.onAssistantTool(p),
-    );
+    // Both tool-trace streams feed the ONE thread: EVENT_ASSISTANT_TOOL from the
+    // voice path, EVENT_CHAT_TOOL from ask_assistant_chat (text). Each chip lands
+    // on the last pending assistant bubble (no backend change).
+    this.unlistenTool = await this.ipc.onAssistantTool((p) => this.onTool(p));
+    this.unlistenChatTool = await this.ipc.onChatTool((p) => this.onTool(p));
   }
 
   /** Release the event subscriptions (e.g. on app teardown). */
@@ -205,18 +213,24 @@ export class AssistantStore {
     this.unlistenListening?.();
     this.unlistenProcessing?.();
     this.unlistenTool?.();
+    this.unlistenChatTool?.();
     this.unlistenWake = null;
     this.unlistenResult = null;
     this.unlistenListening = null;
     this.unlistenProcessing = null;
     this.unlistenTool = null;
+    this.unlistenChatTool = null;
+  }
+
+  /** Empty the conversation (called on each new recording). */
+  clear(): void {
+    this._messages.set([]);
   }
 
   /**
    * Fire the manual "Ask AI" trigger: open the listener AND mark a manual ask in
-   * flight so the assistant card stays visible for the whole round-trip. Errors
-   * (backend rejects / listener unavailable) clear the in-flight flag so the UI
-   * doesn't get stuck pulsing.
+   * flight so the surface stays visible for the whole round-trip. Errors clear
+   * the in-flight flag so the UI doesn't get stuck pulsing.
    */
   async askNow(): Promise<void> {
     this._manualAskInFlight.set(true);
@@ -232,11 +246,8 @@ export class AssistantStore {
   /**
    * CLICK-TO-STOP: stop the open listener so the FULL accumulated utterance is
    * dispatched. Optimistically flip `listening` off + `processing` on so the orb
-   * morphs to PROCESSING the instant the user clicks (the backend's
-   * `{active:false}` listening + `{active:true}` processing events reconcile it
-   * shortly after); the answer clears processing via {@link onResult}. The manual
-   * ask stays in flight so the card keeps the answer's home. A no-op backend
-   * (nothing armed) is fine — `endVoiceCommand` is a graceful no-op there.
+   * morphs to PROCESSING the instant the user clicks; the answer clears processing
+   * via {@link onResult}. A no-op backend (nothing armed) is fine.
    */
   async endAsk(): Promise<void> {
     this._listening.set(false);
@@ -244,7 +255,6 @@ export class AssistantStore {
     try {
       await this.ipc.endVoiceCommand();
     } catch (e) {
-      // The stop call itself failed — don't leave the orb stuck "processing".
       this._processing.set(false);
       this._manualAskInFlight.set(false);
       throw e;
@@ -252,70 +262,122 @@ export class AssistantStore {
   }
 
   /**
-   * Ask the assistant a TYPED question (the text composer). Optimistically prepend
-   * a pending row with the typed text BEFORE the await (the {@link onWake} pattern)
-   * so the user sees it instantly, mark a turn in flight (orb → processing), then
-   * route through the SAME gated brain via IPC. The live tool-trace lands via
-   * {@link onAssistantTool}; the answer resolves the pending row via {@link onResult}.
+   * Send a TEXT message into the unified thread: optimistically append the user
+   * bubble + a pending assistant bubble, ship the FULL clean conversation to the
+   * multi-turn brain (`ask_assistant_chat` → memory), then resolve the assistant
+   * bubble with the reply. The live tool-trace lands via {@link onTool}. A no-op
+   * while another turn is in flight.
    */
-  async askText(text: string): Promise<void> {
-    const command = text.trim();
-    if (!command) return;
-    const row: AssistantInteraction = {
+  async send(text: string): Promise<void> {
+    const t = text.trim();
+    if (!t || this._processing()) return;
+    const userMsg: ChatMessage = {
       id: this.nextId++,
-      source: "text",
-      command,
-      status: "pending",
-      summary: "",
+      role: "user",
+      text: t,
+      status: "ok",
       trace: [],
       citations: [],
     };
-    this._interactions.update((rows) =>
-      [row, ...rows].slice(0, AssistantStore.MAX),
-    );
+    const botMsg: ChatMessage = {
+      id: this.nextId++,
+      role: "assistant",
+      text: "",
+      status: "pending",
+      trace: [],
+      citations: [],
+    };
+    this._messages.update((m) => [...m, userMsg, botMsg]);
     this._processing.set(true);
+
+    // Build a CLEAN conversation payload: every user turn + every real assistant
+    // answer (skip the in-flight bubble + any prior error/empty bubble), newest
+    // user message last — exactly what the backend's format_chat expects.
+    const payload: ChatMsg[] = this._messages()
+      .filter(
+        (m) =>
+          m.role === "user" ||
+          (m.status !== "pending" &&
+            m.status !== "error" &&
+            m.text.trim().length > 0),
+      )
+      .map((m) => ({ role: m.role, text: m.text }));
+
     try {
-      await this.ipc.askAssistantText(command);
-    } catch (e) {
-      this._processing.set(false);
-      // Don't leave the optimistic row stuck "pending" if the dispatch was rejected.
-      this._interactions.update((rows) =>
-        rows.map((r) =>
-          r.id === row.id
-            ? { ...r, status: "error" as const, summary: "Couldn't send your question." }
-            : r,
+      const reply = await this.ipc.askAssistantChat(payload);
+      this._messages.update((m) =>
+        m.map((x) =>
+          x.id === botMsg.id
+            ? {
+                ...x,
+                status: reply.status,
+                text: reply.summary || "(no answer)",
+                citations: parseCitations(reply.citations),
+              }
+            : x,
         ),
       );
-      throw e;
+    } catch {
+      this._messages.update((m) =>
+        m.map((x) =>
+          x.id === botMsg.id
+            ? {
+                ...x,
+                status: "error" as const,
+                text: "Couldn't send your message.",
+              }
+            : x,
+        ),
+      );
+    } finally {
+      this._processing.set(false);
     }
   }
 
+  /**
+   * A wake phrase fired: append a USER bubble (the heard command) + a PENDING
+   * assistant bubble to the SAME thread. The matching {@link onResult} resolves
+   * the assistant bubble.
+   */
   private onWake(p: WakeDetectedPayload): void {
-    const row: AssistantInteraction = {
+    const userMsg: ChatMessage = {
       id: this.nextId++,
-      source: "wake",
-      command: p.command,
-      status: "pending",
-      summary: "",
+      role: "user",
+      text: p.command,
+      status: "ok",
       trace: [],
       citations: [],
     };
-    this._interactions.update((rows) =>
-      [row, ...rows].slice(0, AssistantStore.MAX),
-    );
+    const botMsg: ChatMessage = {
+      id: this.nextId++,
+      role: "assistant",
+      text: "",
+      status: "pending",
+      trace: [],
+      citations: [],
+    };
+    this._messages.update((m) => [...m, userMsg, botMsg]);
   }
 
   /**
-   * Attach a LIVE tool-trace chip to the in-flight (top pending) interaction. A
-   * "running" event pushes a new chip; a "done" event resolves the most recent
-   * matching running chip (or appends one if none). No pending row → ignore (the
-   * trace has no home). Pure immutable signal updates — no NG0600.
+   * Attach a LIVE tool-trace chip to the in-flight (last pending) assistant
+   * bubble. A "running" event pushes a new chip; a "done" event resolves the most
+   * recent matching running chip (or appends one if none). No pending bubble →
+   * ignore (the trace has no home). Pure immutable signal updates — no NG0600.
+   * Shared by BOTH the voice (EVENT_ASSISTANT_TOOL) and text (EVENT_CHAT_TOOL)
+   * tool-trace streams.
    */
-  private onAssistantTool(p: AssistantToolPayload): void {
-    this._interactions.update((rows) => {
-      const idx = rows.findIndex((r) => r.status === "pending");
-      if (idx === -1) return rows;
-      const row = rows[idx];
+  private onTool(p: AssistantToolPayload): void {
+    this._messages.update((msgs) => {
+      let idx = -1;
+      for (let i = msgs.length - 1; i >= 0; i--) {
+        if (msgs[i].role === "assistant" && msgs[i].status === "pending") {
+          idx = i;
+          break;
+        }
+      }
+      if (idx === -1) return msgs;
+      const row = msgs[idx];
       const trace = row.trace.slice();
       if (p.state === "running") {
         trace.push({
@@ -344,7 +406,7 @@ export class AssistantStore {
           });
         }
       }
-      const next = rows.slice();
+      const next = msgs.slice();
       next[idx] = { ...row, trace };
       return next;
     });
@@ -360,41 +422,65 @@ export class AssistantStore {
     if (p.active) this._listening.set(false);
   }
 
+  /**
+   * A voice answer landed. Clear the listening/processing/in-flight state, then
+   * resolve the last pending assistant bubble (summary + citations). If there is
+   * NO pending bubble (a manual "Ask AI" with no preceding wake), append a fresh
+   * USER bubble (the heard command) + a resolved assistant bubble so the voice
+   * turn still lands in the one thread.
+   */
   private onResult(p: VoiceActionResultPayload): void {
-    // The answer landed — the manual ask (if any) is no longer in flight, the
-    // listener is closed, and the dispatch is no longer processing.
     this._manualAskInFlight.set(false);
     this._listening.set(false);
     this._processing.set(false);
-    this._interactions.update((rows) => {
-      // Resolve the most recent still-pending row; if none (a result without a
-      // wake we observed), prepend a fresh resolved row so nothing is lost.
-      const idx = rows.findIndex((r) => r.status === "pending");
+    this._messages.update((msgs) => {
+      let idx = -1;
+      for (let i = msgs.length - 1; i >= 0; i--) {
+        if (msgs[i].role === "assistant" && msgs[i].status === "pending") {
+          idx = i;
+          break;
+        }
+      }
       if (idx === -1) {
-        // A manual ("Ask AI") result has NO preceding wake row — surface the
-        // HEARD command straight from the payload so the card shows what the
-        // user actually said (not an empty "usłyszano: …").
-        const row: AssistantInteraction = {
+        const heard = p.command.trim();
+        const botMsg: ChatMessage = {
           id: this.nextId++,
-          source: "voice",
-          command: p.command,
+          role: "assistant",
+          text: p.summary,
           status: p.status,
-          summary: p.summary,
           trace: [],
           citations: parseCitations(p.citations),
         };
-        return [row, ...rows].slice(0, AssistantStore.MAX);
+        // Only add a user bubble when something was actually heard. A manual ask that caught
+        // NOTHING (empty command, no preceding wake) must not append an empty user turn — it would
+        // render a blank bubble AND ship `{role:"user",text:""}` in the next send()'s history.
+        if (!heard) {
+          return [...msgs, botMsg];
+        }
+        const userMsg: ChatMessage = {
+          id: this.nextId++,
+          role: "user",
+          text: heard,
+          status: "ok",
+          trace: [],
+          citations: [],
+        };
+        return [...msgs, userMsg, botMsg];
       }
-      const next = rows.slice();
+      const next = msgs.slice();
+      // Resolve the assistant bubble.
       next[idx] = {
         ...next[idx],
-        // Keep the wake-detected command, but fall back to the payload's heard
-        // command if the pending row never captured one.
-        command: next[idx].command || p.command,
         status: p.status,
-        summary: p.summary,
+        text: p.summary,
         citations: parseCitations(p.citations),
       };
+      // Backfill the heard command onto the preceding user bubble if the wake
+      // event never captured one (wake fired with an empty trailing command).
+      const prev = next[idx - 1];
+      if (prev && prev.role === "user" && !prev.text.trim() && p.command) {
+        next[idx - 1] = { ...prev, text: p.command };
+      }
       return next;
     });
   }
